Moved CatDatabaseEntity to Models
