+++ app/src/main/java/com/example/catslist/database/CatsDatabase.kt
@@ -5,6 +5,7 @@ import android.util.Log
 import androidx.room.Database
 import androidx.room.Room
 import androidx.room.RoomDatabase
+import com.example.catslist.models.CatDatabaseEntity
 
 @Database(entities = [CatDatabaseEntity::class], version = 2)
 abstract class CatsDatabase : RoomDatabase() {
