+++ app/src/main/java/com/example/catslist/viewmodels/FavoriteCatsListFragmentViewModel.kt
@@ -8,7 +8,7 @@ import android.os.Environment
 import android.widget.Toast
 import androidx.lifecycle.AndroidViewModel
 import com.example.catslist.models.Cat
-import com.example.catslist.database.CatDatabaseEntity
+import com.example.catslist.models.CatDatabaseEntity
 import com.example.catslist.database.CatsDatabaseRepository
 import com.example.catslist.tools.CatStorage
 import kotlinx.coroutines.CoroutineScope
